Using PicklableInferenceSession in the Taherian GMPEs
